+++ plugins/clipboard/plugin.js
@@ -2825,6 +2825,12 @@
 
 					fallbackDataTransfer._isCustomMimeTypeSupported = false;
 
+					// It looks like after our custom MIME type test Edge 17 is denying access on nativeDataTransfer (#2169).
+					// Upstream issue: https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/18089287/
+					if ( CKEDITOR.env.edge && CKEDITOR.env.version >= 17 ) {
+						return true;
+					}
+
 					try {
 						nativeDataTransfer.setData( testType, testValue );
 						fallbackDataTransfer._isCustomMimeTypeSupported = nativeDataTransfer.getData( testType ) === testValue;