+++ tests/plugins/emoji/basic.js
@@ -39,7 +39,7 @@
 			emojiTools.runAfterInstanceReady( editor, bot, function( editor, bot ) {
 				var autocomplete = editor._.emoji.autocomplete;
 
-				bot.setHtmlWithSelection( '<p>foo:da^</p>' );
+				bot.setHtmlWithSelection( '<p>foo :da^</p>' );
 				editor.editable().fire( 'keyup', new CKEDITOR.dom.event( {} ) );
 
 				// Asserted data might not be initialised yet or reset to null value with `onTextUnmatch` method.
@@ -49,7 +49,7 @@
 				// Handle throttle in autocomplete which by defualt is 20ms;
 				setTimeout( function() {
 					resume( function() {
-						bot.setHtmlWithSelection( '<p>foo:dagg^</p>' );
+						bot.setHtmlWithSelection( '<p>foo :dagg^</p>' );
 						editor.editable().fire( 'keyup', new CKEDITOR.dom.event( {} ) );
 						assert.areSame( ':dagg', autocomplete.model.query, 'Model keeps wrong querry.' );
 						assert.areSame( 1, autocomplete.model.data.length, 'Emoji result contains more than one result.' );
@@ -71,7 +71,8 @@
 			// Fallback in case where ajax couldn't load before tests.
 			var data = JSON.stringify( [
 				{ id: ':bug:', symbol: '🐛' },
-				{ id: ':dagger:', symbol: '🗡' }
+				{ id: ':dagger:', symbol: '🗡' },
+				{ id: ':OK_hand:', symbol: '👌' }
 			] );
 
 			stub = sinon.stub( CKEDITOR.ajax, 'load', function( url, callback ) {
@@ -83,6 +84,18 @@
 
 		tearDown: function() {
 			stub.restore();
+
+			for ( var key in this.editorBots ) {
+				var editor = this.editorBots[ key ].editor,
+					autocomplete;
+
+				if ( editor._.emoji ) {
+					autocomplete = editor._.emoji.autocomplete;
+
+					emojiTools.clearAutocompleteModel( autocomplete );
+					autocomplete.close();
+				}
+			}
 		},
 
 		'test emoji objects are added to editor': function( editor ) {
@@ -96,33 +109,86 @@
 			emojiTools.runAfterInstanceReady( editor, bot, function( editor, bot ) {
 				var autocomplete = editor._.emoji.autocomplete;
 
-				bot.setHtmlWithSelection( '<p>foo:bug^</p>' );
+				bot.setHtmlWithSelection( '<p>foo :bug^</p>' );
 				editor.editable().fire( 'keyup', new CKEDITOR.dom.event( {} ) );
 				assert.areSame( ':bug', autocomplete.model.query, 'Model keeps wrong querry.' );
 				assert.areSame( 1, autocomplete.model.data.length, 'Emoji result contains more than one result.' );
 				objectAssert.areEqual( { id: ':bug:', symbol: '🐛' }, autocomplete.model.data[ 0 ], 'Emoji result contains wrong result' );
-
-				emojiTools.clearAutocompleteModel( autocomplete );
-				autocomplete.close();
 			} );
 		},
 
 		'test emoji are not actived when too few letters are written': function( editor, bot ) {
 			emojiTools.runAfterInstanceReady( editor, bot, function( editor, bot ) {
 				var autocomplete = editor._.emoji.autocomplete;
-				bot.setHtmlWithSelection( '<p>foo:b^</p>' );
+				bot.setHtmlWithSelection( '<p>foo :b^</p>' );
 				editor.editable().fire( 'keyup', new CKEDITOR.dom.event( {} ) );
 
 				// Asserted data might not be initialised yet or reset to null value with `onTextUnmatch` method.
 				emojiTools.assertIsNullOrUndefined( autocomplete.model.query );
 				emojiTools.assertIsNullOrUndefined( autocomplete.model.data );
 
-				emojiTools.clearAutocompleteModel( autocomplete );
-				autocomplete.close();
+			} );
+		},
 
+		// (#2167)
+		'test emoji suggestion box is case insensitive': function( editor, bot ) {
+			emojiTools.runAfterInstanceReady( editor, bot, function( editor, bot ) {
+				var autocomplete = editor._.emoji.autocomplete,
+					queries = [ ':OK_HAND', ':ok_hand', ':OK_hand', ':ok_HAND', ':Ok_hanD', 'oK_HANd' ];
+
+				CKEDITOR.tools.array.forEach( queries, function( query, index ) {
+					setTimeout( function() {
+						resume();
+						bot.setHtmlWithSelection( '<p>foo ' + query + '^</p>' );
+						editor.editable().fire( 'keyup', new CKEDITOR.dom.event( {} ) );
+
+						objectAssert.areEqual( { id: ':ok_hand:', symbol: '👌' }, autocomplete.model.data[ 0 ], 'Emoji result contains wrong result' );
+					}, 50 * ( index + 1 ) );
+					wait();
+				} );
+			} );
+		},
+
+		// (#2195)
+		'test emoji suggestion box shouldn\'t appear inside text': function( editor, bot ) {
+			emojiTools.runAfterInstanceReady( editor, bot, function( editor, bot ) {
+				var autocomplete = editor._.emoji.autocomplete,
+					editable = editor.editable();
+
+				bot.setHtmlWithSelection( '<p>foo:bug^</p>' );
+
+				// Delay assertions because of autocomplete throttle.
+				setTimeout( function() {
+					resume( function() {
+						emojiTools.assertIsNullOrUndefined( autocomplete.model.query );
+						emojiTools.assertIsNullOrUndefined( autocomplete.model.data );
+					} );
+				}, 50 );
+
+				editable.fire( 'keyup', new CKEDITOR.dom.event( {} ) );
+				wait();
+			} );
+		},
+
+		// (#2195)
+		'test emoji suggestion box should appear at the beginning of new line': function( editor, bot ) {
+			emojiTools.runAfterInstanceReady( editor, bot, function( editor, bot ) {
+				var autocomplete = editor._.emoji.autocomplete,
+					editable = editor.editable();
+
+				bot.setHtmlWithSelection( '<p>foo</p><p>:bug^</p>' );
+
+				// Delay assertions because of autocomplete throttle.
+				setTimeout( function() {
+					resume( function() {
+						objectAssert.areEqual( { id: ':bug:', symbol: '🐛' }, autocomplete.model.data[ 0 ], 'Emoji result contains wrong result' );
+					} );
+				}, 50 );
+
+				editable.fire( 'keyup', new CKEDITOR.dom.event( {} ) );
+				wait();
 			} );
 		}
-
 	};
 
 	tests = bender.tools.createTestsForEditors( CKEDITOR.tools.objectKeys( bender.editors ), tests );