+++ tests/plugins/image2/init.js
@@ -1,0 +1,36 @@
+/* bender-tags: editor,widget */
+/* bender-include: ../easyimage/_helpers/tools.js */
+/* global easyImageTools */
+
+( function() {
+	'use strict';
+
+	bender.test( {
+		// (#1791)
+		'test plugin init when easyimage is active': function() {
+			if ( easyImageTools.isUnsupportedEnvironment() ) {
+				assert.ignore();
+			}
+
+			var spy = sinon.spy( CKEDITOR, 'warn' );
+
+			bender.editorBot.create( {
+				name: 'easyimage',
+				config: {
+					plugins: 'image2,easyimage,contextmenu'
+				}
+			}, function( bot ) {
+				var editor = bot.editor;
+
+				spy.restore();
+
+				assert.isTrue( spy.calledWith( 'editor-plugin-conflict', { plugin: 'image2', replacedWith: 'easyimage' } ) );
+				assert.isUndefined( editor.commands.image, 'Command' );
+				assert.isUndefined( editor.widgets.registered.image, 'Widget' );
+				assert.isUndefined( editor.getMenuItem( 'image' ), 'Context menu' );
+				assert.isNotUndefined( editor.commands.easyimage, 'Command: Easyimage' );
+			} );
+		}
+	} );
+
+} )();